Wrote script PinCollision and test script BallForward. Updated script SpawnBowlingBall.
Implemented test script BallForward to easily launch a bowling ball forward to collide with pins, testing new script PinCollision which destroys the pins that fall over and the ball after 3 seconds.

SpawnBowlingBall was updated as well which handles respawning of both the bowling ball and the pins. Will be renamed in the future to reflect this.

If the player does a left thumbs up, a bowling ball respawns in front of them, but only one ball can exist at once. This negates any latency that may accidentally allow for a second ball to be spawned if the camera detects a second thumbs up.

Then, there's a check to see if the pins should be reset. This check is if throwCount == 2 and ball.position == pins.position.
throwCount is incremented when SpawnBall() is called and is reset to 0 when SpawnPins() is called.