skeeball game loop, game manager, and accompanying scripts
skeeball game loop is fully implemented with some short changes needed such as the despawn collider properly decrementing ball count. Game manager was implemented to choreograph the logic of the game including score and ball count. SkeeballScore.cs script is used to keep track of the score of the game depending on the collider the skeeball is in contact with.